Expand Systems command evidence negatives
